make a specialized version of debugger for Html programs
