+++ docs/modules/Debug/redux-devtool.ts.md
@@ -1,6 +1,6 @@
 ---
 title: Debug/redux-devtool.ts
-nav_order: 5
+nav_order: 6
 parent: Modules
 ---
