+++ docs/modules/Decode.ts.md
@@ -1,6 +1,6 @@
 ---
 title: Decode.ts
-nav_order: 6
+nav_order: 7
 parent: Modules
 ---
