+++ docs/modules/Html.ts.md
@@ -1,6 +1,6 @@
 ---
 title: Html.ts
-nav_order: 7
+nav_order: 8
 parent: Modules
 ---
