+++ docs/modules/Http.ts.md
@@ -1,6 +1,6 @@
 ---
 title: Http.ts
-nav_order: 8
+nav_order: 9
 parent: Modules
 ---
