+++ docs/modules/Navigation.ts.md
@@ -1,6 +1,6 @@
 ---
 title: Navigation.ts
-nav_order: 10
+nav_order: 11
 parent: Modules
 ---
