+++ docs/modules/Platform.ts.md
@@ -1,6 +1,6 @@
 ---
 title: Platform.ts
-nav_order: 11
+nav_order: 12
 parent: Modules
 ---
