+++ docs/modules/React.ts.md
@@ -1,6 +1,6 @@
 ---
 title: React.ts
-nav_order: 12
+nav_order: 13
 parent: Modules
 ---
