+++ docs/modules/Time.ts.md
@@ -1,6 +1,6 @@
 ---
 title: Time.ts
-nav_order: 15
+nav_order: 16
 parent: Modules
 ---
