+++ docs/modules/index.ts.md
@@ -1,6 +1,6 @@
 ---
 title: index.ts
-nav_order: 9
+nav_order: 10
 parent: Modules
 ---
